aliases/implications: refactor to share approve/reject messages.
